feat: move auth handlers to api
